+++ strongtalk/src/cpp/main/vm/interpreter/Interpreter.cpp
@@ -222,7 +222,7 @@ void Interpreter::wrong_primitive_result() {
 DoubleOop Interpreter::oopify_FloatValue() {
     // Called from float_oopify. Get the float argument by inspecting the stack and the argument of the Floats::oopify operation.
     Frame f = DeltaProcess::active()->last_frame();
-    st_assert( *( f.hp() - 3 ) == static_cast<int>(ByteCodes::Code::float_unary_op_to_oop) and *( f.hp() - 1 ) == static_cast<int>( Floats::Function::oopify ), "not called by Floats::oopify" );
+    st_assert( *( f.hp() - 3 ) == static_cast<std::size_t>(ByteCodes::Code::float_unary_op_to_oop) and *( f.hp() - 1 ) == static_cast<std::size_t>( Floats::Function::oopify ), "not called by Floats::oopify" );
     int float_index = *( f.hp() - 2 );
     st_assert( 0 <= float_index and float_index < max_nof_floats, "illegal float index" );
     double *float_address = (double *) ( (const char *) f.fp() + ( float_0_offset - ( max_nof_floats - 1 ) * SIZEOF_FLOAT ) + float_index * SIZEOF_FLOAT );
