+++ strongtalk/src/cpp/main/vm/interpreter/interpreterStatistics.hpp
@@ -20,8 +20,8 @@ class InterpreterStatistics : AllStatic {
 private:
 
     static bool_t                                                                        _is_initialized;            // true if InterpreterStatistics has been initialized
-    static std::array<std::uint32_t, static_cast<int>(ByteCodes::Code::NUMBER_OF_CODES)> _bytecode_counters;
-    static std::array<int, static_cast<int>(ByteCodes::Code::NUMBER_OF_CODES)>           _bytecode_generation_order;
+    static std::array<std::uint32_t, static_cast<std::size_t>(ByteCodes::Code::NUMBER_OF_CODES)> _bytecode_counters;
+    static std::array<int, static_cast<std::size_t>(ByteCodes::Code::NUMBER_OF_CODES)>           _bytecode_generation_order;
 
     static void reset_bytecode_counters();
